+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Subject label filter no longer incorrectly matches emails that have no subject line.
+
 ## [0.3.0] - 2026-06-01
 
 ### Added